+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.6.2-SNAPSHOT]
 
+- UIComponent.currentWindow property
+- openModal/openWindow defaults to currentWindow as owner (https://github.com/edvin/tornadofx/issues/246)
 - Accordion.fold has `expanded` parameter
 - Fixed: ComboBox with cellFormat does not show bound element (https://github.com/edvin/tornadofx/issues/245)